+++ ShareHubMIU.Domain/Entities/RoomSharing.cs
@@ -9,7 +9,7 @@ namespace ShareHubMIU.Domain.Entities
     public class RoomSharing : Item
     {
         public int Capacity { get; set; }
-        public decimal RentPerMonth { get; set; }
+        public double RentPerMonth { get; set; }
         public bool IsAvailable { get; set; }
         public string RoomType { get; set; } // e.g., Single, Shared
         public bool IsFurnished { get; set; }
